+++ ca/django_ca/tests/extensions/test_admin_html.py
@@ -1,3 +1,4 @@
+"""Test how extensions look like in the admin interface."""
 import typing
 
 from cryptography import x509